+++ auth/build.gradle.kts
@@ -15,7 +15,6 @@
  */
 
 import care.data4life.sdk.auth.dependency.Dependency
-import care.data4life.sdk.auth.LibraryConfig
 
 plugins {
     id("org.jetbrains.kotlin.multiplatform")
@@ -28,7 +27,7 @@ plugins {
     id("care.data4life.sdk.auth.publishing-config")
 }
 
-group = LibraryConfig.group
+group = care.data4life.gradle.auth.config.LibraryConfig.group
 
 kotlin {
     android {
@@ -131,11 +130,11 @@ kotlin {
 }
 
 android {
-    compileSdkVersion(LibraryConfig.android.compileSdkVersion)
+    compileSdkVersion(care.data4life.gradle.auth.config.LibraryConfig.android.compileSdkVersion)
 
     defaultConfig {
-        minSdkVersion(LibraryConfig.android.minSdkVersion)
-        targetSdkVersion(LibraryConfig.android.targetSdkVersion)
+        minSdkVersion(care.data4life.gradle.auth.config.LibraryConfig.android.minSdkVersion)
+        targetSdkVersion(care.data4life.gradle.auth.config.LibraryConfig.android.targetSdkVersion)
 
         versionCode = 1
         versionName = "${project.version}"
@@ -154,7 +153,7 @@ android {
         }
     }
 
-    resourcePrefix(LibraryConfig.android.resourcePrefix)
+    resourcePrefix(care.data4life.gradle.auth.config.LibraryConfig.android.resourcePrefix)
 
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_1_8
@@ -180,14 +179,14 @@ android {
     }
 }
 
-val d4lClientConfig = care.data4life.sdk.auth.D4LConfigHelper.loadClientConfigAndroid("$rootDir")
+val d4lClientConfig = care.data4life.gradle.auth.D4LConfigHelper.loadClientConfigAndroid("$rootDir")
 val jvmAssetsPath = "${projectDir}/src/jvmTest/resources"
 
 val provideTestConfig: Task by tasks.creating {
     doLast {
         val asset = File(jvmAssetsPath)
         if (!asset.exists()) asset.mkdirs()
-        File(jvmAssetsPath, "client_config.json").writeText(care.data4life.sdk.auth.D4LConfigHelper.toJson(d4lClientConfig))
+        File(jvmAssetsPath, "client_config.json").writeText(care.data4life.gradle.auth.D4LConfigHelper.toJson(d4lClientConfig))
     }
 }
