+++ buildSrc/src/main/kotlin/care/data4life/gradle/auth/config/LibraryConfig.kt
@@ -14,7 +14,7 @@
  * contact D4L by email to help@data4life.care.
  */
 
-package care.data4life.sdk.auth
+package care.data4life.gradle.auth.config
 
 object LibraryConfig {
